feat(kyc): add first name field to KYC result payload and refactor notification logic
- Introduced a new optional `first_name` field in the `ActivepiecesKycResultPayload` type for personalized email greetings.
- Refactored the KYC result notification logic to utilize the updated payload structure across various KYC processing routes.
- Updated utility functions to ensure consistent handling of KYC result notifications, enhancing user experience with personalized communication.

These changes improve the clarity and personalization of KYC notifications, ensuring users receive relevant updates based on their verification outcomes.
